Add rudimentary configuration file support
The configuration file is stored in ~/.morc, and is in ini format
with three sections: [core], [sort], and [tag].

The options are named after the internal configuration variables,
rather than being parsed as arguments, so (for example) --file-name
and all the --prefix-foo arguments are set using 'file-name=...' in
the config file.

Command line options override options given in the file.

Mine:

[sort]
safe-paths: windows
file-name: {disc?}{track!d:02d} - {title}